Check in form exists and goes to data form
Data form is automatically populated with as much data as can be gotten
and appropriately filled in from the drchrono api and previous
check-ins. Pressing submit on the data form will eventually save the
check-in such that doctors may see it and all the associated data.

Besides that, the only things that need thought are how to prevent
patients from just navigating backwards and seeing everyone's social
security number, and also how to give doctors a live feed of check-ins
rather than have them need to refresh the page to see new check-ins