Session locked out of sensitive data after entering kiosk
To be allowed back into the home page, the browser must be restarted
and the doctor must log in.

This works perfectly on firefox, but chrome seems to be bad at deleting
browser length cookies when the browser it shut. What this means is that
on chrome, not even shutting the browser down and restarting it will let
the doctor go to the home page, however the doctor can get around that by
opening an incognito window. That case shouldn't come up often, as the
kiosk computer isn't intended to be used for looking at check-in information
anyways, that should happen on a separate machine in parallel to the kiosk
running